add linux mac hash
