+++ include/vcpkg/base/system.mac.h
@@ -1,0 +1,10 @@
+#include <vcpkg/base/optional.h>
+#include <vcpkg/base/stringview.h>
+
+#include <string>
+
+namespace vcpkg
+{
+    Optional<StringView> find_first_nonzero_mac(StringView sv);
+    std::string get_user_mac();
+}
