+++ src/vcpkg/metrics.cpp
@@ -3,6 +3,7 @@
 #include <vcpkg/base/hash.h>
 #include <vcpkg/base/json.h>
 #include <vcpkg/base/strings.h>
+#include <vcpkg/base/system.mac.h>
 #include <vcpkg/base/system.process.h>
 #include <vcpkg/base/uuid.h>
 
@@ -19,67 +20,6 @@
 namespace vcpkg
 {
     LockGuarded<Metrics> g_metrics;
-
-    Optional<StringView> find_first_nonzero_mac(StringView sv)
-    {
-        static constexpr StringLiteral ZERO_MAC = "00-00-00-00-00-00";
-
-        auto first = sv.begin();
-        const auto last = sv.end();
-
-        while (first != last)
-        {
-            // XX-XX-XX-XX-XX-XX
-            // 1  2  3  4  5  6
-            // size = 6 * 2 + 5 = 17
-            first = std::find_if(first, last, ParserBase::is_hex_digit);
-            if (last - first < 17)
-            {
-                break;
-            }
-
-            bool is_first = true;
-            bool is_valid = true;
-            auto end_of_mac = first;
-            for (int i = 0; is_valid && i < 6; ++i)
-            {
-                if (!is_first)
-                {
-                    if (*end_of_mac != '-')
-                    {
-                        is_valid = false;
-                        break;
-                    }
-                    ++end_of_mac;
-                }
-                is_first = false;
-
-                if (!ParserBase::is_hex_digit(*end_of_mac))
-                {
-                    is_valid = false;
-                    break;
-                }
-                ++end_of_mac;
-
-                if (!ParserBase::is_hex_digit(*end_of_mac))
-                {
-                    is_valid = false;
-                    break;
-                }
-                ++end_of_mac;
-            }
-            if (is_valid && StringView{first, end_of_mac} != ZERO_MAC)
-            {
-                return StringView{first, end_of_mac};
-            }
-            else
-            {
-                first = end_of_mac;
-            }
-        }
-
-        return nullopt;
-    }
 
     static std::string get_current_date_time_string()
     {
@@ -246,30 +186,6 @@ namespace vcpkg
     static bool g_should_print_metrics = false;
     static std::atomic<bool> g_metrics_disabled = true;
 
-#if defined(_WIN32)
-    static std::string get_MAC_user()
-    {
-        if (!LockGuardPtr<Metrics>(g_metrics)->metrics_enabled())
-        {
-            return "{}";
-        }
-
-        auto getmac = cmd_execute_and_capture_output(Command("getmac"));
-
-        if (getmac.exit_code != 0) return "0";
-
-        auto found_mac = find_first_nonzero_mac(getmac.output);
-        if (auto p = found_mac.get())
-        {
-            return Hash::get_string_hash(*p, Hash::Algorithm::Sha256);
-        }
-        else
-        {
-            return "0";
-        }
-    }
-#endif
-
     void Metrics::set_send_metrics(bool should_send_metrics) { g_should_send_metrics = should_send_metrics; }
 
     void Metrics::set_print_metrics(bool should_print_metrics) { g_should_print_metrics = should_print_metrics; }
@@ -298,13 +214,11 @@ namespace vcpkg
             write_config = true;
         }
 
-#if defined(_WIN32)
         if (config.user_mac.empty())
         {
-            config.user_mac = get_MAC_user();
+            config.user_mac = get_user_mac();
             write_config = true;
         }
-#endif
 
         if (write_config)
         {
